# Speed up CLI startup by loading SecureRandom only when needed