@kke 2017-05-24
Not needed if 1.2.3 is buried. The more complex version targeted to 1.3.0 in #2335 already includes this.